Create README.md
Disclaimer and Credits: I strive to respect the intellectual property of others and give appropriate credit where due. If you notice any code, ideas, or resources in this repository that require proper attribution, please feel free to contact me or open an issue. I will promptly review and update the repository with the necessary acknowledgments.